HIA-1277: Add feature flag to send decimals

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/assessRisksAndNeeds/ArnRiskOfSeriousRecidivismScore.kt
@@ -1,10 +1,11 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.assessRisksAndNeeds
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.RiskOfSeriousRecidivism
+import java.math.BigDecimal
 
 data class ArnRiskOfSeriousRecidivismScore(
   val scoreLevel: String? = null,
-  val percentageScore: Int? = null,
+  val percentageScore: BigDecimal? = null,
 ) {
   fun toRiskOfSeriousRecidivism(useV2NumericalValue: Boolean = true): RiskOfSeriousRecidivism =
     RiskOfSeriousRecidivism(
